Fixes and changes

+++ Host/main.cpp
@@ -1,6 +1,6 @@
 #include "Host.h"
 
-int main(int argc, char* argv[]) {
+int main(size_t argc, char* argv[]) {
 	UI This;
 	This.loop(argc, argv);
 	return 0;
